% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleaning.R
\name{cleanData}
\alias{cleanData}
\title{Cleans the data frame}
\usage{
cleanData(df, removeNA, removeDuplicates, cleanNames)
}
\arguments{
\item{df}{a dataframe to be cleaned}

\item{removeNA}{a boolean to remove rows with NA values}

\item{removeDuplicates}{a boolean to remove duplicate rows}

\item{cleanNames}{a boolean to clean the column names using the janitor package}
}
\value{
A cleaned dataframe
}
\description{
Cleans the data frame
}
