% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleaning.R
\name{cleanImport}
\alias{cleanImport}
\title{Load in a file and clean it}
\usage{
cleanImport(
  fileName,
  sheetName = 1,
  headerRow = TRUE,
  removeNA = TRUE,
  removeDuplicates = TRUE,
  cleanNames = TRUE
)
}
\arguments{
\item{fileName}{The name of the file to be imported}

\item{sheetName}{The name of the sheet to be opened (only for xls and xlsx files)}

\item{headerRow}{a boolean if the file contains a header row (only for txt files)}

\item{removeNA}{a boolean to remove rows with NA values}

\item{removeDuplicates}{a boolean to remove duplicate rows}

\item{cleanNames}{a boolean to clean the column names using the janitor package}
}
\value{
A cleaned data frame
}
\description{
cleanImport() returns a cleaned data frame given a path (file name). It supports file types including csv, txt, xls, and xlsx. It decides the appropriate read.x function to use based on the path name and cleans the data according to user specifications.
}
\references{
\url{https://i.imgur.com/7yBm8wY.jpg}
}
