% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAnalyzer-package.R
\docType{package}
\name{DataAnalyzer-package}
\alias{DataAnalyzer}
\alias{DataAnalyzer-package}
\title{DataAnalyzer: Probability and Statistics R 'Shiny' Modules}
\description{
Repository for different 'shiny' modules that are specialized in in analyzing data using Probability and Statistics methods. Analysis output is presented via attractive `shiny` output and interactive `cavasXpress` \url{https://cran.r-project.org/package=canvasXpress} plots.
}
\author{
\strong{Maintainer}: Mohammed Ali \email{mohammed@aggregate-genius.com}

}
\keyword{internal}
