% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarizeData}
\alias{summarizeData}
\title{Return a summary of two variables of a given data frame}
\usage{
summarizeData(dataFrame, var1, var2)
}
\arguments{
\item{dataFrame}{The name of the data frame}

\item{var1}{A quoted variable name from the data frame}

\item{var2}{A second quoted variable name from the data frame}
}
\value{
A vector of tables containing the summaries of the variables
}
\description{
Return a summary of two variables of a given data frame
}
