% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleaning.R
\name{cleanData}
\alias{cleanData}
\title{Clean a data frame}
\usage{
cleanData(df, removeNA, removeDuplicates, cleanNames)
}
\arguments{
\item{df}{a data frame to be cleaned}

\item{removeNA}{a boolean to remove rows with NA values}

\item{removeDuplicates}{a boolean to remove duplicate rows}

\item{cleanNames}{a boolean to clean the column names using the janitor package}
}
\value{
A cleaned data frame
}
\description{
cleanData executes common data cleaning functions according to a few function inputs.
}
