% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleaning.R
\name{cleanData}
\alias{cleanData}
\title{Cleans the data in the data frame}
\usage{
cleanData(filename, removeNA)
}
\arguments{
\item{removeNA}{a boolean to remove rows with NA values}

\item{dataframe}{a dataframe to be cleaned}
}
\value{
A dataframe
}
\description{
Cleans the data in the data frame
}
