% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleaning.R
\name{cleanImport}
\alias{cleanImport}
\title{Loads a file given a format, with optional treatment of NA values}
\usage{
cleanImport(filename, filetype, removeNA = TRUE)
}
\arguments{
\item{filename}{The name of the file to be imported}

\item{filetype}{The type of the file ("csv" = (.csv), "excel" = (.xls, .xlsx) "text" = (txt))}

\item{removeNA}{a boolean to remove rows with NA values}
}
\value{
A dataframe
}
\description{
Loads a file given a format, with optional treatment of NA values
}
