% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleaning.R
\name{cleanImport}
\alias{cleanImport}
\title{Loads in a file and cleans it according to user specification, returning the result in a dataframe}
\usage{
cleanImport(
  fileName,
  sheetName = 1,
  headerRow = TRUE,
  removeNA = TRUE,
  removeDuplicates = TRUE,
  cleanNames = TRUE
)
}
\arguments{
\item{fileName}{The name of the file to be imported}

\item{sheetName}{The name of the sheet to be opened (only for xls and xlsx files)}

\item{headerRow}{a boolean if the file contains a header row (only for txt files)}

\item{removeNA}{a boolean to remove rows with NA values}

\item{removeDuplicates}{a boolean to remove duplicate rows}

\item{cleanNames}{a boolean to clean the column names using the janitor package}
}
\value{
A cleaned dataframe
}
\description{
Loads in a file and cleans it according to user specification, returning the result in a dataframe
}
