% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cleaning.R
\name{cleanImport}
\alias{cleanImport}
\title{Loads a file given a format, with optional treatment of NA values}
\usage{
cleanImport(
  file,
  type,
  removeNA = TRUE,
  removeDuplicates = TRUE,
  cleanNames = TRUE
)
}
\arguments{
\item{file}{The type of the file ("csv" = (.csv), "excel" = (.xls, .xlsx) "text" = (txt))}

\item{removeNA}{a boolean to remove rows with NA values}

\item{removeDuplicates}{a boolean to remove duplicate rows}

\item{cleanNames}{a boolean to clean the column names using the janitor package}
}
\value{
A cleaned dataframe
}
\description{
Loads a file given a format, with optional treatment of NA values
}
